+++ back-end/api/quickstart/test_posts.py
@@ -5,7 +5,7 @@ from rest_framework import status
 from django.test import TestCase, Client
 from .models import Author, Post
 from .serializers import AuthorSerializer, PostSerializer
-from .helpers_test import get_test_post_fields
+from .helper_test import get_test_post_fields
 
 client = Client()
